+++ src/com/amplitude/api/DeviceInfo.java
@@ -372,11 +372,8 @@ public class DeviceInfo {
             Location location = null;
             try {
                 location = locationManager.getLastKnownLocation(provider);
-            } catch (IllegalArgumentException e) {
-                // failed to get last known location from provider
-                Diagnostics.getLogger().logError("Failed to get most recent location", e);
-            } catch (SecurityException e) {
-                // failed to get last known location from provider
+            } catch (Exception e) {
+                AmplitudeLog.getLogger().w(TAG, "Failed to get most recent location");
                 Diagnostics.getLogger().logError("Failed to get most recent location", e);
             }
             if (location != null) {